reflect refactored store
