+++ src/app/appStore.ts
@@ -17,4 +17,4 @@ export type AppStore = typeof store
 // Infer the `AppDispatch` type from the store itself
 export type AppDispatch = typeof store.dispatch
 // Same for the `RootState` type
-export type RootState = ReturnType<typeof store.getState>
+export type AppRootState = ReturnType<typeof store.getState>
