+++ src/app/dataStore.ts
@@ -1,5 +1,5 @@
 import { configureStore } from '@reduxjs/toolkit'
-import featuresReducer from './features/geoFeatures/geoFeaturesSlice'
+import featuresReducer from '../features/geoFeatures/geoFeaturesSlice'
 
 export const dataStore = configureStore({
   reducer: {
