+++ src/components/Layers.tsx
@@ -5,8 +5,8 @@ import './Layers.css';
 import { LineChartOutlined } from '@ant-design/icons';
 import { Feature } from 'geojson'
 import { REFERENCE_POINT_TYPE, TRACK_TYPE, ZONE_TYPE } from '../constants';
-import { useDataDispatch, useDataSelector } from '../app/hooks';
-import { selectedFeaturesSelection, SelectionState } from '../features/selection/selectionSlice';
+import { useAppSelector, useDataDispatch, useDataSelector } from '../app/hooks';
+import { SelectionState } from '../features/selection/selectionSlice';
 
 interface LayerProps {
   openGraph: {(): void}
@@ -14,8 +14,8 @@ interface LayerProps {
 
 const Layers: React.FC<LayerProps> = ({openGraph}) => {
   const features = useDataSelector(state => state.featureCollection.features)
-  const selectedFeatureIds = useDataSelector(state => state.selected.selected)
-  const selectedFeatures = useDataSelector(selectedFeaturesSelection)
+  const selectedFeatureIds = useAppSelector(state => state.selected.selected)
+  const selectedFeatures = features.filter((feature) => selectedFeatureIds.includes(feature.id as string))
   const dispatch = useDataDispatch()
 
   const [model, setModel] = React.useState<TreeDataNode[]>([])
