+++ src/components/Map.tsx
@@ -1,7 +1,7 @@
 import { Feature, Geometry, MultiPoint } from "geojson";
 import { MapContainer, Marker, Popup, GeoJSON, TileLayer, CircleMarker as ReactCircleMarker } from 'react-leaflet'
 import { PathOptions, StyleFunction, LatLngExpression, CircleMarker, LeafletMouseEvent } from 'leaflet'
-import { useDataSelector, useAppDispatch } from "../app/hooks";
+import { useDataSelector, useAppDispatch, useAppSelector } from "../app/hooks";
 import { TRACK_TYPE, ZONE_TYPE } from "../constants";
 import Track from "./Track";
 import Zone from "./Zone";
@@ -48,7 +48,8 @@ const calcInterpLocation = (poly: MultiPoint, times: any, current: number, index
 
 const Map: React.FC = () => {
   const features = useDataSelector(state => state.featureCollection.features)
-  const selectedFeaturesId = useAppSelector(state => state.selected.selected)
+  const selectedFeatureIds = useAppSelector(state => state.selected.selected)
+
   const {current} = useAppSelector(state => state.time)
   const dispatch = useAppDispatch();
 
@@ -59,7 +60,7 @@ const Map: React.FC = () => {
       if (feat?.properties?.color) {
         res.color = feat.properties.color
       }
-      if(selectedFeaturesId.includes(feature.id as string)) {
+      if(selectedFeatureIds.includes(feature.id as string)) {
         res.color = '#aaa'
       }
     }
@@ -84,7 +85,7 @@ const Map: React.FC = () => {
   const onClickHandler = useCallback((id: string, modifier: boolean): void => {
     if (modifier) {
       // add/remove from selection
-      if (selectedFeaturesId.includes(id)) {
+      if (selectedFeatureIds.includes(id)) {
         dispatch({type: 'selection/removeSelection', payload: id as string})
       } else {
         dispatch({type: 'selection/addSelection', payload: id as string})
@@ -93,7 +94,7 @@ const Map: React.FC = () => {
       // just select this item
       dispatch({type: 'selection/selectionChanged', payload: {selected: [id as string]}})
     }
-  }, [dispatch, selectedFeaturesId])
+  }, [dispatch, selectedFeatureIds])
 
   const onTooltipClick = useCallback( (event: LeafletMouseEvent) => {
     if (event.target.feature) {
