+++ src/components/Properties.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { Table, Tooltip } from 'antd';
-import { useDataSelector } from '../app/hooks';
+import { useAppSelector, useDataSelector } from '../app/hooks';
 import './Properties.css';
 
 const formatItem = (value: any) => {
@@ -23,7 +23,7 @@ const formatItem = (value: any) => {
 
 const Properties: React.FC = () => {
   const allFeatures = useDataSelector(state => state.featureCollection.features)
-  const selectedFeatureIds= useDataSelector(state => state.selected.selected)
+  const selectedFeatureIds= useAppSelector(state => state.selected.selected)
   const features = allFeatures.filter(feature => selectedFeatureIds.includes(feature.id as string))
 
   if (!features || features.length === 0) {
