+++ src/main.tsx
@@ -3,8 +3,8 @@ import { createRoot } from 'react-dom/client'
 import './index.css'
 import App from './App.tsx'
 import { Provider } from 'react-redux'
-import { store } from './app/store'
-import { dataStore } from './dataStore'
+import { store } from './app/appStore.ts'
+import { dataStore } from './app/dataStore.ts'
 
 createRoot(document.getElementById('root')!).render(
   <StrictMode>
